Seperated row and column flex
